Add new module StronglinkSL032TokenProvider
This adds a token provider interface implementation for the
Stronglink SL032 RFID reader. It is a straight-forward, proof-of-concept
implementation for now, based on the gensio library.
(This allows to have the RFID reader connected physically to another host.)

The software architecture could be improved later, but it works so far
for the moment.

Signed-off-by: Michael Heimpold <mhei@heimpold.de>